workflow auth controller
