+++ ptychodus/controller/workflow/__init__.py
@@ -1,0 +1,5 @@
+from .controller import WorkflowController
+
+__all__ = [
+    'WorkflowController',
+]
